v1.6.6

+++ src/Service/Service_Bill.js
@@ -1,4 +1,5 @@
 const { Bill } = require("../Model/Bill");
+const { Storage } = require("../Model/Storage");
 const { Subscription } = require("../Model/Subscription");
 const { User } = require("../Model/User");
 const { Create_Id } = require("../Util/Create_Id");
@@ -68,24 +69,32 @@ const SV__Get_Bill__Current = (User_Id, Sub_Id) => {
         });
       }
 
-      const checkSub = Subscription.findOne({ Sub_Id });
-      if (!checkSub) {
+      if (Sub_Id) {
+        const checkSub = Subscription.findOne({ Sub_Id });
+        if (!checkSub) {
+          return resolve({
+            status: 404,
+            message: "Not found subscription!",
+            data: {},
+            error: "Not found subscription!",
+          });
+        }
+
+        const result = await Bill.findOne({ User_Id, Sub_Id });
+        if (!result) {
+          return resolve({
+            status: 404,
+            message: "Not found bill for user",
+          });
+        }
         return resolve({
-          status: 404,
-          message: "Not found subscription!",
-          data: {},
-          error: "Not found subscription!",
+          status: 200,
+          message: "get bill complete!",
+          data: result,
         });
       }
 
-      const result = await Bill.findOne({ User_Id, Sub_Id });
-      if (!result) {
-        return resolve({
-          status: 404,
-          message: "Not found bill for user",
-        });
-      }
-
+      const result = await Bill.find({ User_Id });
       resolve({
         status: 200,
         message: "get all bill complete!",
@@ -180,10 +189,43 @@ const SV__Create_Bill = (User_Id, Sub_Id) => {
 const SV__Delete_Bill = (Bill_Id) => {
   return new Promise(async (resolve, reject) => {
     try {
+      const getBill = await Bill.findOne({ Bill_Id });
+      if (!getBill) {
+        return resolve({ status: 404, message: "Not found Bill with id" });
+      }
+
       const result = await Bill.findOneAndDelete({ Bill_Id });
       if (!result) {
         return resolve({ status: 404, message: "Not found Bill with id" });
       }
+
+      const getSub = await Subscription.findOne({ Sub_Id: getBill.Sub_Id });
+      if (!getSub) {
+        return resolve({
+          status: 404,
+          message: "Not found subscription with id",
+        });
+      }
+
+      const getStorage = await Storage.findOne({ User_Id: getBill.User_Id });
+      if (!getStorage) {
+        return resolve({
+          status: 404,
+          message: "Not found subscription with id",
+        });
+      }
+      const UpdateStorage = await Storage.findOneAndUpdate(
+        { User_Id: getBill.User_Id },
+        { Limit: getStorage.Limit - getSub.Storage },
+        { new: true }
+      );
+      if (!UpdateStorage) {
+        return resolve({
+          status: 404,
+          message: "Not found subscription with id",
+        });
+      }
+
       resolve({
         status: 200,
         message: "Delete Bill complete!",
